Moved cache clear funcitonality to update-transaction.js
